https://sonarcloud.io/project/issues?id=codewars-solutions&issues=AXAV3OoXe9SUkepG-zTr&open=AXAV3OoXe9SUkepG-zTr

+++ kata/6 kyu/playing-with-digits/main/DigPow.java
@@ -1,8 +1,8 @@
 import static java.util.stream.IntStream.range;
 
 class DigPow {
-  static long digPow(int n, int p) {
-    int sum = range(0, (n + "").length()).map(i -> (int) Math.pow((n + "").charAt(i) - 48, i + p)).sum();
+  static long digPow(int n, double p) {
+    int sum = range(0, (n + "").length()).map(i -> (int) Math.pow((n + "").charAt(i) - 48., i + p)).sum();
     return sum % n == 0 ? sum / n : -1;
   }
 }
